accountmicroservice/src/main/java/com/example/accountmicroservice/controller/MovementController.java
package com.example.accountmicroservice.controller;

import com.example.accountmicroservice.models.MovementDTO;
import com.example.accountmicroservice.service.MovementService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/movimientos")
public class MovementController {
    private final MovementService movementService;

    @Autowired
    public MovementController(MovementService movementService) {
        this.movementService = movementService;
    }

    @GetMapping
    public ResponseEntity<List<MovementDTO>> getAllMovements() {
        return ResponseEntity.ok(movementService.getAllMovements());
    }

    @GetMapping("/{id}")
    public ResponseEntity<MovementDTO> getMovementById(@PathVariable Long id) {
        MovementDTO mouvement = movementService.getMovementById(id);
        return ResponseEntity.ok(mouvement);
    }

    @PostMapping
    public ResponseEntity<MovementDTO> createMovement(@RequestBody MovementDTO movementDTO) {
        return ResponseEntity.status(HttpStatus.CREATED).body(movementService.createMovement(movementDTO));
    }

    @PutMapping("/{id}")
    public ResponseEntity<MovementDTO> updateMovement(@PathVariable Long id, @RequestBody MovementDTO movementDTO) {
        return ResponseEntity.ok(movementService.updateMovement(id, movementDTO));
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteMovement(@PathVariable Long id) {
        movementService.deleteMovement(id);
        return ResponseEntity.noContent().build();
    }
}
